feat: 카드사 선택 컴포넌트 및 관련 상수 추가, 카드 브랜드 결정 로직 수정
