+++ my-app/src/components/Card/Card.tsx
@@ -5,7 +5,7 @@ import visaLogo from "../../assets/visaLogo.png";
 import amexLogo from "../../assets/amexLogo.png";
 import unionpayLogo from "../../assets/unionpayLogo.png";
 import dinersLogo from "../../assets/dinersLogo.png";
-import { decideCardColor } from "../../utils/decideCardInfo";
+import { decideCardColor } from "../../constants/cardCompanies";
 
 const LOGO_MAP: Record<string, string> = {
   master: masterLogo,
